pyproject.toml
[project]
name = "q-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "huggingface-hub>=1.2.3",
    "llama-cpp-python>=0.3.16",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "sqlite-utils>=3.39",
    "uvicorn[standard]>=0.38.0",
]
